Removed menu items, modified mobile layout.
