+++ website/src/pages/cwkw/cwkw-layout.tsx
@@ -2,18 +2,32 @@ import "@fontsource/quattrocento-sans/latin.css"
 import "normalize.css"
 import React from "react"
 import { Helmet } from "react-helmet"
+import { HiOutlineBookOpen } from "react-icons/hi"
+import { MdMenu } from "react-icons/md"
+import {
+  Dialog,
+  DialogBackdrop,
+  DialogDisclosure,
+  useDialogState,
+} from "reakit/Dialog"
 import { Link } from "src/components"
+import { useMediaQuery } from "src/custom-hooks"
 import Footer from "src/footer"
+import { navButton } from "src/menu.css"
+import { mediaQueries } from "src/style/constants"
 import "src/style/global.css"
-import { MobileNav, NavMenu } from "../../menu"
 import { HeaderPrefDrawer } from "../../mode"
 import { PreferencesProvider } from "../../preferences-context"
 import "../../wordpress.css"
 import * as css from "./cwkw-layout.css"
 import { themeClass } from "./theme.css"
+import { Chapters } from "./toc-sidebar"
+import * as tocCss from "./toc-sidebar.css"
 
 /** Wrapper for cwkw site pages, providing them with a navigation header and footer. */
 const CWKWLayout: React.FC = ({ children }) => {
+  const isDesktop = useMediaQuery(mediaQueries.large)
+
   return (
     <PreferencesProvider>
       <Helmet titleTemplate="%s - DAILP" defaultTitle="DAILP">
@@ -24,7 +38,7 @@ const CWKWLayout: React.FC = ({ children }) => {
       </Helmet>
       <header aria-label="Site Header" id="header" className={css.header}>
         <div className={css.headerContents}>
-          <MobileNav menuID={171} />
+          {!isDesktop && <MobileTOC />}
           <div className={css.contentContainer}>
             <h1 className={css.siteTitle}>
               <Link className={css.siteLink} href="/">
@@ -37,11 +51,42 @@ const CWKWLayout: React.FC = ({ children }) => {
           </div>
           <HeaderPrefDrawer />
         </div>
-        <NavMenu menuID={171} />
       </header>
       {children}
       <Footer />
     </PreferencesProvider>
+  )
+}
+
+const MobileTOC = () => {
+  const dialog = useDialogState({ animated: true })
+
+  return (
+    <>
+      <DialogBackdrop {...dialog} className={tocCss.tocDrawerBg}>
+        <Dialog
+          {...dialog}
+          className={tocCss.tocDrawer}
+          as="nav"
+          aria-label="Table of Contents"
+          preventBodyScroll={true}
+        >
+          <nav className={tocCss.sidebar}>
+            <div className={tocCss.tocHeader}>
+              <div>Table of Contents</div>
+            </div>
+            <Chapters />
+          </nav>
+        </Dialog>
+      </DialogBackdrop>
+      <DialogDisclosure
+        {...dialog}
+        className={navButton}
+        aria-label="Open Table of Contents Drawer"
+      >
+        <MdMenu size={32} />
+      </DialogDisclosure>
+    </>
   )
 }
